// src/app/interfaces/domain.ts
export class domain {
    domain: string | undefined;
    usedStorage: string | undefined;
    monthlyVisitor: number | undefined;
    availableDomains: number | undefined;
    domainTag: string | undefined;
    status: string | undefined;
    usedDomains: number | undefined;
    planName: string | undefined;
    subdomain: any;
}